Move .timeseries.Sample to .core.Sample
This is mainly to break a dependency cycle between actors and timeseries
utilities. Any actor that work on timeseries needs to import Sample,
which lived in .timeseries, but .timeseries also import all utilities
working on timeseries (like resampler and logical meter) that in turn
use some actors.

Moving Sample to a (really) leaf module solves the issue.
