+++ src/frequenz/sdk/timeseries/__init__.py
@@ -5,19 +5,17 @@
 Handling of timeseries streams.
 
 A timeseries is a stream (normally an async iterator) of
-[samples][frequenz.sdk.timeseries.Sample].
+[samples][frequenz.sdk.core.Sample].
 
 This module provides tools to operate on timeseries.
 """
 
 from ._logical_meter import LogicalMeter
 from ._resampler import GroupResampler, Resampler, ResamplingFunction
-from ._sample import Sample
 
 __all__ = [
     "GroupResampler",
     "LogicalMeter",
     "Resampler",
     "ResamplingFunction",
-    "Sample",
 ]
