+++ src/frequenz/sdk/timeseries/_logical_meter/_logical_meter.py
@@ -11,8 +11,8 @@ from frequenz.channels import Broadcast, Receiver, Sender
 
 from ...actor._channel_registry import ChannelRegistry
 from ...actor._data_sourcing.microgrid_api_source import ComponentMetricRequest
+from ...core import Sample
 from ...microgrid.component import ComponentMetricId
-from .._sample import Sample
 from ._formula_builder import FormulaBuilder
 from ._formula_engine import FormulaEngine
